Remove FC, change radar file structure and naming
