--- src/shared/components/radar/radarContent/index.ts
@@ -1,1 +1,0 @@
-export { RadarContent } from './radarContent.component';
